+++ R/dependencies.R
@@ -1,4 +1,6 @@
 
+`_renv_dependencies` <- new.env(parent = emptyenv())
+
 #' Find R Package Dependencies in a Project
 #'
 #' Find \R packages used within a project.
@@ -40,6 +42,8 @@
 #' /data/
 #' ```
 #'
+#' @inheritParams renv-params
+#'
 #' @param path The path to a (possibly multi-mode) \R file, or a directory
 #'   containing such files. By default, all files within the current working
 #'   directory are checked, recursively.
@@ -49,8 +53,11 @@
 #'   explicitly to ensure that your project's `.renvignore`s (if any) are
 #'   properly handled.
 #'
-#' @param quiet Boolean; report problems discovered (if any) during dependency
-#'   discovery?
+#' @param progress Boolean; report progress output while enumerating
+#'   dependencies?
+#'
+#' @param errors How should errors that occur during dependency enumeration be
+#'   handled? See **Errors** for more details.
 #'
 #' @param dev Boolean; include 'development' dependencies as well? That is,
 #'   packages which may be required during development but are unlikely to be
@@ -59,6 +66,26 @@
 #'
 #' @return An \R `data.frame` of discovered dependencies, mapping inferred
 #'   package names to the files in which they were discovered.
+#'
+#' @section Errors:
+#'
+#' `renv`'s attempts to enumerate package dependencies in your project can fail
+#' -- most commonly, because of parse errors in your \R code. The `errors`
+#' parameter can be used to control how `renv` responds to errors that occur.
+#'
+#' \tabular{ll}{
+#' **Name** \tab **Action** \cr
+#' `"reported"` \tab Errors are reported to the user, but are otherwise ignored. \cr
+#' `"fatal"`    \tab Errors are fatal and stop execution. \cr
+#' `"ignored"`  \tab Errors are ignored and not reported to the user. \cr
+#' }
+#'
+#' Depending on the structure of your project, you may want `renv` to ignore
+#' errors that occur when attempting to enumerate dependencies. However, a more
+#' robust solution would be to use an `.renvignore` file to tell `renv` not to
+#' scan such files for dependencies, or to configure the project to require
+#' explicit dependency management (`renv::settings$snapshot.type("explicit")`)
+#' and enumerate your dependencies in a project `DESCRIPTION` file.
 #'
 #' @section Development Dependencies:
 #'
@@ -77,25 +104,17 @@
 #' renv::dependencies()
 #'
 #' }
-dependencies <- function(path = getwd(),
-                         root = NULL,
-                         quiet = FALSE,
-                         dev = FALSE)
+dependencies <- function(
+  path = getwd(),
+  root = NULL,
+  ...,
+  progress = TRUE,
+  errors = c("reported", "fatal", "ignored"),
+  dev = FALSE)
 {
   renv_scope_error_handler()
 
-  path <- renv_path_normalize(path, winslash = "/", mustWork = TRUE)
-  root <- root %||% renv_dependencies_root(path)
-
-  renv_dependencies_begin(root = root)
-  on.exit(renv_dependencies_end(), add = TRUE)
-
-  files <- renv_dependencies_find(path, root)
-  deps <- renv_dependencies_discover(files, quiet)
-
-  if (!quiet)
-    renv_dependencies_report()
-
+  deps <- delegate(renv_dependencies_impl)
   if (empty(deps) || nrow(deps) == 0L)
     return(deps)
 
@@ -103,7 +122,30 @@ dependencies <- function(path = getwd(),
     deps <- deps[!deps$Dev, ]
 
   deps
+}
 
+renv_dependencies_impl <- function(
+  path = getwd(),
+  root = NULL,
+  ...,
+  progress = TRUE,
+  errors = c("reported", "fatal", "ignored"),
+  dev = FALSE)
+{
+  path <- renv_path_normalize(path, winslash = "/", mustWork = TRUE)
+  root <- root %||% renv_dependencies_root(path)
+
+  if (exists(path, envir = `_renv_dependencies`))
+    return(get(path, envir = `_renv_dependencies`))
+
+  renv_dependencies_begin(root = root)
+  on.exit(renv_dependencies_end(), add = TRUE)
+
+  files <- renv_dependencies_find(path, root)
+  deps <- renv_dependencies_discover(files, progress, errors)
+  renv_dependencies_report(errors)
+
+  deps
 }
 
 renv_dependencies_root <- function(path = getwd()) {
@@ -217,16 +259,14 @@ renv_dependencies_find_dir_children <- function(path, root) {
 
 }
 
-renv_dependencies_discover <- function(paths, quiet) {
+renv_dependencies_discover <- function(paths, progress, errors) {
 
-  if (!renv_dependencies_discover_preflight(paths, quiet))
+  if (!renv_dependencies_discover_preflight(paths, errors))
     return(invisible(NULL))
 
-  # short path if we're being quiet
-  if (quiet || renv_testing()) {
-    deps <- lapply(paths, renv_dependencies_discover_impl)
-    return(bind_list(deps))
-  }
+  # short path if we're not showing progress
+  if (identical(progress, FALSE))
+    return(bapply(paths, renv_dependencies_discover_impl))
 
   # otherwise, run with progress reporting
 
@@ -243,10 +283,6 @@ renv_dependencies_discover <- function(paths, quiet) {
 
 renv_dependencies_discover_impl <- function(path) {
 
-  entry <- renv_filebacked_get("dependencies", path)
-  if (!is.null(entry))
-    return(entry)
-
   callback <- renv_dependencies_callback(path)
   if (is.null(callback)) {
     fmt <- "internal error: no callback registered for file %s"
@@ -257,15 +293,17 @@ renv_dependencies_discover_impl <- function(path) {
   if (inherits(result, "condition"))
     return(NULL)
 
-  renv_filebacked_set("dependencies", path, result)
   result
 
 }
 
-renv_dependencies_discover_preflight <- function(paths, quiet) {
+renv_dependencies_discover_preflight <- function(paths, errors) {
+
+  if (identical(errors, "ignored"))
+    return(TRUE)
 
   limit <- renv_config("dependencies.limit", default = 1000L)
-  if (quiet || length(paths) < limit)
+  if (length(paths) < limit)
     return(TRUE)
 
   lines <- c(
@@ -277,6 +315,9 @@ renv_dependencies_discover_preflight <- function(paths, quiet) {
     ""
   )
   vwritef(lines, length(paths))
+
+  if (identical(errors, "reported"))
+    return(TRUE)
 
   if (interactive() && !proceed()) {
     message("* Operation aborted.")
@@ -1002,7 +1043,10 @@ renv_dependencies_error <- function(path, error = NULL, packages = NULL) {
 
 }
 
-renv_dependencies_report <- function() {
+renv_dependencies_report <- function(errors) {
+
+  if (identical(errors, "ignored"))
+    return(FALSE)
 
   state <- renv_dependencies_state()
   if (is.null(state))
@@ -1012,7 +1056,7 @@ renv_dependencies_report <- function() {
   if (empty(problems))
     return(TRUE)
 
-  vwritef("WARNING: One or more problems were discovered while discovering dependencies.\n", con = stderr())
+  ewritef("WARNING: One or more problems were discovered while enumerating dependencies.\n")
 
   # bind into list
   bound <- bapply(problems, function(problem) {
@@ -1031,9 +1075,47 @@ renv_dependencies_report <- function() {
     prefix <- format(paste("ERROR", seq_along(problem$message)))
     messages <- paste(prefix, problem$message, sep = ": ", collapse = "\n\n")
     text <- c(file, lines, "", messages, "")
-    vwritef(text, con = stderr())
+    ewritef(text)
   })
 
+  ewritef("Please see `?renv::dependencies` for more information.")
+
+  if (identical(errors, "fatal")) {
+    fmt <- "one or more errors occurred while enumerating dependencies"
+    stopf(fmt)
+  }
+
+  renv_condition_signal("renv.dependencies.error", problems)
   TRUE
 
+}
+
+renv_dependencies_scope <- function(path, action, .envir = NULL) {
+
+  path <- renv_path_normalize(path, winslash = "/", mustWork = TRUE)
+  if (exists(path, envir = `_renv_dependencies`))
+    return(get(path, envir = `_renv_dependencies`))
+
+  errors <- renv_config("dependency.errors", default = "reported")
+  message <- paste(action, "aborted")
+
+  deps <- withCallingHandlers(
+    dependencies(path, progress = FALSE, errors = errors, dev = TRUE),
+    renv.dependencies.error = renv_dependencies_error_handler(message, errors)
+  )
+
+  assign(path, deps, envir = `_renv_dependencies`)
+
+  envir <- .envir %||% parent.frame()
+  defer(rm(list = path, envir = `_renv_dependencies`), envir = envir)
+
+}
+
+renv_dependencies_error_handler <- function(message, errors) {
+  list(message, errors)
+  function(condition) {
+    if (identical(errors, "fatal") || interactive() && !proceed())
+      stop(message)
+    renv_condition_data(condition)
+  }
 }