+++ R/errors.R
@@ -32,17 +32,34 @@ renv_error_format_srcref <- function(call, srcref) {
 
 renv_error_simplify <- function(object) {
 
-  longexpr <-
+  case(
+    is.function(object)  ~ renv_error_simplify_function(object),
+    is.recursive(object) ~ renv_error_simplify_recursive(object),
+    TRUE                 ~ object
+  )
+
+}
+
+renv_error_simplify_function <- function(object) {
+  f <- function() {}
+  formals(f) <- formals(object)
+  body(f) <- quote({ ... })
+  f
+}
+
+renv_error_simplify_recursive <- function(object) {
+
+  longcall <-
     is.call(object) &&
     is.symbol(object[[1]]) &&
     object[[1]] == as.name("{") &&
     length(object) >= 8
 
-  if (longexpr)
+  if (longcall)
     return(quote(...))
 
   for (i in seq_along(object))
-    if (is.call(object[[i]]))
+    if (!is.null(object[[i]]))
       object[[i]] <- renv_error_simplify(object[[i]])
 
   object
@@ -60,6 +77,9 @@ renv_error_format <- function(calls, frames) {
       if (!inherits(formatted, "error"))
         return(formatted)
     }
+
+    if (is.function(call[[1]]))
+      return("<condition-handler>(...)")
 
     format(renv_error_simplify(call))
 
@@ -79,12 +99,12 @@ renv_error_format <- function(calls, frames) {
     paste(prefix, code, sep = "")
   })
 
-  header <- "Traceback (most recent calls first):"
+  header <- "Traceback (most recent calls last):"
   c(header, annotated)
 
 }
 
-renv_error_handler <- function(...) {
+renv_error_handler <- function() {
   calls <- head(sys.calls(), n = -1L)
   frames <- head(sys.frames(), n = -1L)
   formatted <- renv_error_format(calls, frames)