+++ R/restore.R
@@ -111,10 +111,10 @@ restore <- function(project  = NULL,
   }
 
   # get records for R packages currently installed
-  current <- snapshot(project = project,
-                      library = library,
+  current <- snapshot(project  = project,
+                      library  = library,
                       lockfile = NULL,
-                      type = "all")
+                      type     = "all")
 
   # compare lockfile vs. currently-installed packages
   diff <- renv_lockfile_diff_packages(current, lockfile)