+++ tests/testthat/test-retrieve.R
@@ -49,7 +49,7 @@ test_that("packages with an unknown source are retrieved from CRAN", {
 test_that("we can retrieve packages from Bitbucket", {
 
   skip_on_cran()
-  skip_on_ci()  # TODO: test fails periodically
+  skip("unreliable test")
 
   record <- list(
     Package        = "skeleton",